refactor(cli): split file validation into multiple modules
Move validation logic into dedicated files for better maintainability.
Initialize the core validation package once at the top level and
reuse it across functions instead of passing it or re-declaring it.
