+++ packages/cli/src/validators/index.ts
@@ -1,0 +1,2 @@
+export * from './slideAutoFix.js';
+export * from './slideValidator.js';
